pyproject.toml
[project]
name = "waiswallet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notebook>=7.5.3",
    "pandas>=3.0.0",
    "sqlalchemy>=2.0.46",
    "python-dotenv>=1.0.1",
    "pydantic-ai>=1.51.0",
    "fastapi>=0.128.0",
    "httpx>=0.27.0",
    "ipykernel>=7.1.0",
]
